ServiceList: Reintegrated ColumnFilterDropDown into ServiceList, added search text highlighting and auto-focus of search field
